Crafting in inventory

+++ Outer Space/Outer Space/Outer Space/Level.cs
@@ -152,15 +152,15 @@ namespace Outer_Space
                     int random = Globals.Randomizer.Next(0, 3);
                     if (random == 0)
                     {
-                        Rewards.Add(new Weapon(Player, Globals.Randomizer.Next(0, Weapon.ListOfMethods().Count()), EnemyDifficulty));
+                        Rewards.Add(new Weapon(Player, Globals.Randomizer.Next(0, Weapon.ListOfMethods().Count()), (int)EnemyDifficulty));
                     }
                     else if (random == 1)
                     {
-                        Rewards.Add(new Shield(new Vector2(200, Globals.ScreenSize.Y - 30), 100, 10, 100 + (int)EnemyDifficulty * 20 + Globals.Randomizer.Next(0, 30), Globals.Randomizer.Next(0, Shield.ListOfShieldMethods().Count()), EnemyDifficulty));
+                        Rewards.Add(new Shield(new Vector2(200, Globals.ScreenSize.Y - 30), 100, 10, 100 + (int)EnemyDifficulty * 20 + Globals.Randomizer.Next(0, 30), Globals.Randomizer.Next(0, Shield.ListOfShieldMethods().Count()), (int)EnemyDifficulty));
                     }
                     if (random == 2)
                     {
-                        Rewards.Add(new Hull(Player, Globals.Randomizer.Next(0, Hull.ListOfHullMethods().Count()), EnemyDifficulty));
+                        Rewards.Add(new Hull(Player, Globals.Randomizer.Next(0, Hull.ListOfHullMethods().Count()), (int)EnemyDifficulty));
                     }
 
                     // Consumable
